pyproject.toml
[project]
name = "brawlstar-agent"
version = "0.1.0"
description = "Brawl Stars AI research sandbox — capture, perception, and offline analysis"
requires-python = ">=3.12"
dependencies = [
    "opencv-python",
    "numpy",
    "Pillow",
    "yt-dlp",
    "pytesseract",
    "jupyter",
    "matplotlib",
    "httpx",
    "python-dotenv",
]

[tool.uv]
cache-dir = "/media/lin/disk2/.uv-cache"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/brawlstar_agent"]
